backend/requirements.txt
aniso8601==8.0.0
backcall==0.2.0
click==7.1.2
decorator==4.4.2
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy==0.782
mypy-extensions==0.4.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pymongo==3.10.1
python-dotenv==0.13.0
pytz==2020.1
rope==0.17.0
six==1.15.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
wcwidth==0.2.5
Werkzeug==1.0.1
